backend/requirements.txt
django==1.10.4
djangorestframework==3.5.3
django-cors-headers==1.3.1
pillow==3.4.2
hashids==1.1.0
django-appconf==1.0.2
django-imagekit==3.3
pilkit==1.1.13
django-shell-plus
ipython
django-extensions
